renamed ConductStyles file to conductStyles
